[test] Extract properties
