+++ tests/src/given-component-associative-header.cpp
@@ -4,7 +4,7 @@
  * SPDX - License - Identifier: MIT
  */
 
-#include "associative.hpp"
+#include "properties/associative.hpp"
 
 #include "pbdt/bdd.hpp"
