+++ tests/src/runnable-scenario-model-header.cpp
@@ -4,7 +4,7 @@
  * SPDX - License - Identifier: MIT
  */
 
-#include "model.hpp"
+#include "properties/model.hpp"
 
 #include "pbdt/bdd.hpp"
