+++ tests/src/then-component-idempotent-module.cpp
@@ -4,7 +4,7 @@
  * SPDX - License - Identifier: MIT
  */
 
-#include "idempotent.hpp"
+#include "properties/idempotent.hpp"
 
 import exstd;
 import pbdt;
